@@ -1,8 +1,17 @@
-import { writable } from 'svelte/store';
 import type { FileNode } from './types';
 
-export const currentPath = writable<string>('/');
-export const files = writable<FileNode[]>([]);
-export const selectedFiles = writable<Set<string>>(new Set());
-export const viewMode = writable<'grid' | 'list'>('grid');
-export const isLoading = writable<boolean>(false);
+// Global state using Svelte 5 runes pattern
+// These are used as shared state across components
+
+export function createFileManagerState() {
+  return {
+    currentPath: $state<string>('/'),
+    files: $state<FileNode[]>([]),
+    selectedFiles: $state<Set<string>>(new Set()),
+    viewMode: $state<'grid' | 'list'>('grid'),
+    isLoading: $state<boolean>(false),
+  };
+}
+
+// Type for the state
+export type FileManagerState = ReturnType<typeof createFileManagerState>;
