@@ -11,10 +11,10 @@ const config = {
     alias: {
       '$lib': './src/lib',
       '$lib/*': './src/lib/*'
-    },
-    csrf: {
-      checkOrigin: false
     }
+  },
+  compilerOptions: {
+    runes: true
   }
 };
 
