@@ -4,9 +4,6 @@ import { defineConfig } from 'vite';
 export default defineConfig({
   plugins: [sveltekit()],
   server: {
-    port: 3000,
-    fs: {
-      allow: ['..']
-    }
+    port: 3000
   }
 });
